[*] Change join command return data #89
