+++ src/zappy_server_src/core/include/rcodes.h
@@ -38,7 +38,7 @@ static const retcodes_t retcodes[] = {
     {.repr = "202 Logout succeed as : %s\n",
     .code = 202,
     .params = NULL},
-    {.repr = "211 Successufly joined the game in the team : %s\n",
+    {.repr = "211 Successufly joined the game : %s\n",
     .code = 211,
     .params = NULL},
     {.repr = "212 Inventory content : %s\n",
